Build for 2.10
*Something* changed about end-of-input handling in 2.10.
I'm not sure what, but as a result LexerReader now includes
one EOF token that is NOT marked "atEnd" before the infinite
stream of atEnd EOF commences.

The other big change is introducing IndexedSeqShim to cover
the signature difference of TraversableOnce's toIndexedSeq
method between 2.9 and 2.10.

And finally, a small change: introducing a "protecet implicit
def selfContext = this" into the DatasetContext classes, in order
to make creating ColumnNames within their bodies easier.