Modify soql to default to sorting nulls last when ascending, and nulls first when descending for a more performant mapping to how postgres does things.
There are a few possible issues with this:

1. It potentially could break compatibility with API users.

2. It is a very postgres driven decision and may not be right for some future
store.

A slightly different approach would be to change the null placement to be an
Option and let each secondary decide how to order most efficiently... but then
you are exposing functional differences between secondaries and the scope cascades
into the serialization format.

Based on all that, this seems like a sane approach given current realities.