Massively simplify type checking
SoQLAnalysisType is no longer a thing!  I hated SoQLAnalysisType!

The resulting variance of Function and MonomorphicFunction is gone, and
Function's constraints are now sets of SoQLTypes instead of functions
from Any to Boolean.

Type inference is no longer ambiguous -- if there is a valid checking,
some checking will be found.  Not sure I like this yet!