Table alias rework
Before: table aliases were declared with `AS identifier` where
"identifier" had to conform to the lexical rules for identifier-ness.

Now: table aliases can still be declared that way, but they can
also be declared with `AS @tablename`.  This has two advantages:

1. It makes declaration reflect use.
2. Lexically it allows 4x4s to be used for table identifiers.

Note that even before, `@tablename` wasn't necessarily always a 4x4,
thanks to resource names.

The upshot of all this is that we can expand `JOIN @some-thng` where
"some-thng" is a derived view or other structure that isn't just a
bare table into `JOIN (select ...) AS @some-thng` and have it work.

Also this makes the alias on table-functions optional, as we can use
the function-name as the alias reliably now.