EN-55111: Add more rollup hints (#258)
This new rollup application capability is going to require hint to activate. It is safer and allow us to experiment and learn with less worry (not unlike ff). This also has other benefits like not forcing us to incur unnecessary overhead in looking up rollups especially when we haven't optimized rollup discovery.