Add a "remove trivial joins" rewrite pass.
Currently, the only way to do an unanchored query from a subquery is to
query `FROM @single_row` and join to the query you're actually interested
in `ON true`.  This is required for soql-syntax reasons we can't change
(yet) but it's just extra work afterward.  This introduces a rewrite pass
that eliminates these needless extra queries.