Delay typechecking of ON clauses
This lets you use select list references in them, as long as the select-
list column you're referencing only itself references tables that have
been established at the point the ON column exists.

There's still an unfortunate thing here because it would also be nice to
do this same sort of thing for LATERAL joins, but that would be a much
more major restructuring, since the output of the FROM-clause's
typechecking is used as the input to typechecking the select-list, so
it becomes circular.  This ON clause change was easy because they don't
affect the query's input schema at all, just which rows exist.