Optionally sloppy UDF params in the new-analyzer, for OA compatibility
The old-analyzer doesn't know about UDFs at all, so core just inserts
a cast to make sure that the values are the expected type.  This has
some surprising impacts, in that the system incorrectly accepts UDF calls
with mal-typed parameters, as any type errors get deferred to runtime.

This adds an option to do the same thing in the NA.  Let us hope that
we never have to turn it on.