EN-79270: Remove bouncycastle dependency
We want to depend on a different version in other systems, and it
was causing conflicts.  Since all we use it for is :id and :version
obfuscation, just bring in enough code to handle that ourselves.

On the plus side, I get to fulfill a long-standing wish re: not
converting things to byte arrays!