Start to despecialize "this"
With this commit, "this" can be referred to more than once in a query,
and it always refers to the current contextual statement.

What this does _not_ do is change the "required" check to be better
than requiring a "no" or "yes, yes" answer to "do you have a from, and if
so is it 'from this'?"  It also does not change it so that "this" is not
required to have an alias, which is a grammar thing and thus impacts
the old-analyzer codepath.